Converting to int

+++ dm_generator/GenerateSampleAdjacency.py
@@ -64,7 +64,7 @@ class GenerateSampleAdjacency(object):
         hdf5_handle = h5py.File(filename, "w")
 
         for resolution in resolutions:
-            local_size = np.ceil(d_size / resolution)
+            local_size = int(np.ceil(d_size / resolution))
             print(resolution, d_size, local_size)
             d_sample = np.zeros([local_size, local_size], dtype='int32')
             d_sample += self.create_matrix(local_size)
